// song-play-counts/models/top-n-song.ts
export interface TopNSong {
    songid?: number;
    songname?: string;
    artist?: string;
    writer?: string;
    album?: string;
    releaseyear?: number;
    yearPlayCount?: number;
    january?: number;
    february?: number;
    march?: number;
    april?: number;
    may?: number;
    june?: number;
    july?: number;
    august?: number;
    september?: number;
    october?: number;
    november?: number;
    december?: number;
}

export default TopNSong;